Correctly normalize parameters for oAuth authentication
Currently oAuth authenfication for nested arrays is broken as those
information is not decoded like the clients encode the information.
Nested array handling is not covered at all.
This patch simplifies the normalization handling and processes
nested arrays in a proper way. Unfortunately oAuth documentation
fails to give some expectations on how to normalize nested array
information correctly, so each dimension is sorted on its own.